Fix service not stopping on system shutdown/restart
Apparently services need to opt in to being notified if it is being stopped due to system shutdown.

I would've thought the OnStop method would have been called on shutdown, but my other project (as of this commit, still not released to public yet) has revealed this isn't the case.