BREAKING CHANGE: Full Blast config changes
This changes Full Blast to flip a bit (or bits) specified in the config, instead of overwriting the whole register.

This is a bit safer than just blindly writing a constant value to that register, as it seems to store other unknown values that differ depending on the laptop.

However, this does require a config change, which is why it is marked as a BREAKING CHANGE!