The beginning of Hotkey Handler, take two
- Start the hotkey binding UI over from scratch

- Add basic hotkey parsing

- Remove MSI Center hotkey and settings (may be added back later if more reliable MSI Center key detection method is found)

Hotkeys do not function yet and there is still much to be implemented/fixed. A debug label that shows whether a hotkey binding is in progress has been left in in this commit.

Feel free to report bugs with the hotkey binding UI, however there are a few known bugs:

- Some key names may not match expected names (e.g. numbers show as "D0"-"D9", many special characters show up as Oem characters...)

- The "Switch to fan profile" parsing isn't properly implemented yet and results in a crash when saving the hotkey config