Add config for MSI Vector 16 HX AI A2XWIG
Added on behalf of @Peter42R.

This laptop in particular is special: it is the first to require temperature down thresholds stored as absolute values (every other MSI laptop before this one stored down thresholds as offsets from the up thresholds). Changes were made to YAMDCC a few weeks ago to accomodate this, but no config existed for these laptops until now.